+++ src/config.rs
@@ -1,5 +1,6 @@
 //! # Key-value configuration management.
 
+use std::env;
 use std::str::FromStr;
 
 use anyhow::{ensure, Context as _, Result};
@@ -317,6 +318,11 @@ impl Context {
 
     /// Get a configuration key. Returns `None` if no value is set, and no default value found.
     pub async fn get_config(&self, key: Config) -> Result<Option<String>> {
+        let env_key = format!("DELTACHAT_{}", key.as_ref().to_uppercase());
+        if let Ok(value) = env::var(env_key) {
+            return Ok(Some(value));
+        }
+
         let value = match key {
             Config::Selfavatar => {
                 let rel_path = self.sql.get_raw_config(key.as_ref()).await?;
@@ -418,7 +424,7 @@ impl Context {
         match key {
             Config::Selfavatar => {
                 self.sql
-                    .execute("UPDATE contacts SET selfavatar_sent=0;", paramsv![])
+                    .execute("UPDATE contacts SET selfavatar_sent=0;", ())
                     .await?;
                 match value {
                     Some(value) => {