+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 ## Unreleased
 
 ### Changes
+- Drop unused SQL columns #4141
+- "full message view" not needed because of footers that go to contact status #4151
+- Pick up system's light/dark mode in generated message HTML #4150
+- Support non-persistent configuration with DELTACHAT_* env
+
+### Fixes
+- Fix segmentation fault if `dc_context_unref()` is called during
+  background process spawned by `dc_configure()` or `dc_imex()`
+  or `dc_jsonrpc_instance_t` is unreferenced
+  during handling the JSON-RPC request. #4153
+- Delete expired messages using multiple SQL requests. #4158
+- Do not emit "Failed to run incremental vacuum" warnings on success. #4160
 - Ability to send backup over network and QR code to setup second device #4007
 
 ## 1.111.0
@@ -21,6 +33,7 @@
 ### API-Changes
 - jsonrpc: add more advanced API to send a message. #4097
 - jsonrpc: add get webxdc blob API `getWebxdcBlob` #4070
+
 
 ## 1.110.0